Crop controller: ajout put et del Crop service: ajout put et del Location controller: modification du get Location service: ajout unit: retrait des times unit controller: ajout unit service: ajout user controller: modification pour service user service: ajout vegetable controller: netoyage et changement nom fonction vegetable service: netoyage

+++ src/main/java/com/restapi/potatrocapi/Service/CropService.java
@@ -1,18 +1,20 @@
 package com.restapi.potatrocapi.Service;
 
-        import java.util.Date;
-        import java.util.List;
-        import java.util.Optional;
+    import java.util.Date;
+    import java.util.List;
+    import java.util.Optional;
 
-        import com.restapi.potatrocapi.model.Crop;
-        import com.restapi.potatrocapi.model.User;
-        import com.restapi.potatrocapi.model.Vegetable;
-        import com.restapi.potatrocapi.repository.CropRepository;
-        import com.restapi.potatrocapi.repository.UserRepository;
-        import com.restapi.potatrocapi.repository.VegetableRepository;
-        import com.restapi.potatrocapi.tools.tokenInfo;
-        import org.springframework.beans.factory.annotation.Autowired;
-        import org.springframework.stereotype.Service;
+    import com.restapi.potatrocapi.model.Crop;
+    import com.restapi.potatrocapi.model.User;
+    import com.restapi.potatrocapi.model.Vegetable;
+    import com.restapi.potatrocapi.repository.CropRepository;
+    import com.restapi.potatrocapi.repository.UserRepository;
+    import com.restapi.potatrocapi.repository.VegetableRepository;
+    import com.restapi.potatrocapi.tools.tokenInfo;
+    import org.springframework.beans.factory.annotation.Autowired;
+    import org.springframework.http.HttpStatus;
+    import org.springframework.http.ResponseEntity;
+    import org.springframework.stereotype.Service;
 
 
 @Service
@@ -21,7 +23,6 @@ public class CropService {
     CropRepository cropRepository;
     @Autowired
     VegetableRepository vegetableRepository;
-
     @Autowired
     private UserRepository userRepository;
 
@@ -35,7 +36,6 @@ public class CropService {
         } else {
             return null;
         }
-
     }
 
     public List<Crop> getAllCropForUser() {
@@ -57,11 +57,36 @@ public class CropService {
         return cropRepository.save(crop);
     }
 
-    public Crop editCrop(Crop crop) {
-        return cropRepository.save(crop);
+    public ResponseEntity<Crop> editCrop(long id, Crop crop) {
+        Optional<Crop> cropData = cropRepository.findById(id);
+        if(cropData.isPresent()) {
+            String authid = new tokenInfo().getUserSub();
+
+            User user = userRepository.findByAuthid(authid);
+            Vegetable vegetable = crop.getVegetable();
+            vegetable = vegetableRepository.findById(vegetable.getId()).get();
+
+            Crop _crop = cropData.get();
+            _crop.setUpdatedAt(new Date());
+            _crop.setUser(user);
+            _crop.setvegetable(vegetable);
+            _crop.setQuantity(crop.getQuantity());
+            return new ResponseEntity<>(cropRepository.save(_crop), HttpStatus.OK);
+        }
+        else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
-    public void deleteCrop(long id) {
-        cropRepository.deleteById(id);
+
+    public ResponseEntity<String> deleteCrop(long id) {
+        Optional<Crop> cropData = cropRepository.findById(id);
+        if(cropData.isPresent()) {
+            cropRepository.deleteById(id);
+            return new ResponseEntity<>("Crop has been deleted!", HttpStatus.OK);
+        }
+        else {
+            return new ResponseEntity<>("Crop don't exist!",HttpStatus.NOT_FOUND);
+        }
     }
 }
